Refactor Searchbar+Backdrop Added
