+++ content/src/scripts/components/Backdrop/index.js
@@ -1,0 +1,1 @@
+export { default } from "./Backdrop";
