+++ content/src/scripts/containers/SearchBar/index.js
